// cluster/portal/portal/web/package/src/apis/userv1/userv1.ts
export interface ListServiceOptions {
    /**
     * @generated from protobuf field: octelium.api.main.meta.v1.CommonListOptions common = 1;
     */
    common?: CommonListOptions;
    /**
     * @generated from protobuf field: string namespace = 2;
     */
    namespace: string;
}
/**
 * @generated from protobuf message octelium.api.main.user.v1.ListNamespaceOptions
 */
export interface ListNamespaceOptions {
    /**
     * @generated from protobuf field: octelium.api.main.meta.v1.CommonListOptions common = 1;
     */
    common?: CommonListOptions;
}
/**
 * @generated from protobuf message octelium.api.main.user.v1.Namespace
 */
export interface Namespace {
    /**
     * @generated from protobuf field: string apiVersion = 1;
     */
    apiVersion: string;
    /**
     * @generated from protobuf field: string kind = 2;
     */
    kind: string;
    /**
     * Metadata is the object's metadata.
     *
     * @generated from protobuf field: octelium.api.main.meta.v1.Metadata metadata = 3;
     */
    metadata?: Metadata;
    /**
     * Spec is the Namespace specification.
     *
     * @generated from protobuf field: octelium.api.main.user.v1.Namespace.Spec spec = 4;
     */
    spec?: Namespace_Spec;
    /**
     * Status is the current status of the Namespace.
     *
     * @generated from protobuf field: octelium.api.main.user.v1.Namespace.Status status = 5;
     */
    status?: Namespace_Status;
}
/**
 * @generated from protobuf message octelium.api.main.user.v1.Namespace.Spec
 */
export interface Namespace_Spec {
}
/**
 * @generated from protobuf message octelium.api.main.user.v1.Namespace.Status
 */
export interface Namespace_Status {
}
/**
 * @generated from protobuf message octelium.api.main.user.v1.NamespaceList
 */
export interface NamespaceList {
    /**
     * APIVersion is the API version of the object.
     *
     * @generated from protobuf field: string apiVersion = 1;
     */
    apiVersion: string;
    /**
     * Kind is the resource name (i.e. `NamespaceList`).
     *
     * @generated from protobuf field: string kind = 2;
     */
    kind: string;
    /**
     * Items is the list of Namespaces.
     *
     * @generated from protobuf field: repeated octelium.api.main.user.v1.Namespace items = 3;
     */
    items: Namespace[];
    /**
     * ListResponseMeta is common information about the list.
     *
     * @generated from protobuf field: octelium.api.main.meta.v1.ListResponseMeta listResponseMeta = 4;
     */
    listResponseMeta?: ListResponseMeta;
}
// ...
class ListServiceOptions$Type extends MessageType<ListServiceOptions> {
    constructor() {
        super("octelium.api.main.user.v1.ListServiceOptions", [
            { no: 1, name: "common", kind: "message", T: () => CommonListOptions },
            { no: 2, name: "namespace", kind: "scalar", T: 9 /*ScalarType.STRING*/ }
        ]);
    }
    create(value?: PartialMessage<ListServiceOptions>): ListServiceOptions {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.namespace = "";
        if (value !== undefined)
            reflectionMergePartial<ListServiceOptions>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: ListServiceOptions): ListServiceOptions {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* octelium.api.main.meta.v1.CommonListOptions common */ 1:
                    message.common = CommonListOptions.internalBinaryRead(reader, reader.uint32(), options, message.common);
                    break;
                case /* string namespace */ 2:
                    message.namespace = reader.string();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: ListServiceOptions, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* octelium.api.main.meta.v1.CommonListOptions common = 1; */
        if (message.common)
            CommonListOptions.internalBinaryWrite(message.common, writer.tag(1, WireType.LengthDelimited).fork(), options).join();
        /* string namespace = 2; */
        if (message.namespace !== "")
            writer.tag(2, WireType.LengthDelimited).string(message.namespace);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message octelium.api.main.user.v1.ListServiceOptions
 */
export const ListServiceOptions = new ListServiceOptions$Type();
// @generated message type with reflection information, may provide speed optimized methods
class ListNamespaceOptions$Type extends MessageType<ListNamespaceOptions> {
    constructor() {
        super("octelium.api.main.user.v1.ListNamespaceOptions", [
            { no: 1, name: "common", kind: "message", T: () => CommonListOptions }
        ]);
    }
    create(value?: PartialMessage<ListNamespaceOptions>): ListNamespaceOptions {
        const message = globalThis.Object.create((this.messagePrototype!));
        if (value !== undefined)
            reflectionMergePartial<ListNamespaceOptions>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: ListNamespaceOptions): ListNamespaceOptions {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* octelium.api.main.meta.v1.CommonListOptions common */ 1:
                    message.common = CommonListOptions.internalBinaryRead(reader, reader.uint32(), options, message.common);
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: ListNamespaceOptions, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* octelium.api.main.meta.v1.CommonListOptions common = 1; */
        if (message.common)
            CommonListOptions.internalBinaryWrite(message.common, writer.tag(1, WireType.LengthDelimited).fork(), options).join();
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message octelium.api.main.user.v1.ListNamespaceOptions
 */
export const ListNamespaceOptions = new ListNamespaceOptions$Type();
// @generated message type with reflection information, may provide speed optimized methods
class Namespace$Type extends MessageType<Namespace> {
    constructor() {
        super("octelium.api.main.user.v1.Namespace", [
            { no: 1, name: "apiVersion", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 2, name: "kind", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 3, name: "metadata", kind: "message", T: () => Metadata },
            { no: 4, name: "spec", kind: "message", T: () => Namespace_Spec },
            { no: 5, name: "status", kind: "message", T: () => Namespace_Status }
        ]);
    }
    create(value?: PartialMessage<Namespace>): Namespace {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.apiVersion = "";
        message.kind = "";
        if (value !== undefined)
            reflectionMergePartial<Namespace>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: Namespace): Namespace {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* string apiVersion */ 1:
                    message.apiVersion = reader.string();
                    break;
                case /* string kind */ 2:
                    message.kind = reader.string();
                    break;
                case /* octelium.api.main.meta.v1.Metadata metadata */ 3:
                    message.metadata = Metadata.internalBinaryRead(reader, reader.uint32(), options, message.metadata);
                    break;
                case /* octelium.api.main.user.v1.Namespace.Spec spec */ 4:
                    message.spec = Namespace_Spec.internalBinaryRead(reader, reader.uint32(), options, message.spec);
                    break;
                case /* octelium.api.main.user.v1.Namespace.Status status */ 5:
                    message.status = Namespace_Status.internalBinaryRead(reader, reader.uint32(), options, message.status);
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: Namespace, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* string apiVersion = 1; */
        if (message.apiVersion !== "")
            writer.tag(1, WireType.LengthDelimited).string(message.apiVersion);
        /* string kind = 2; */
        if (message.kind !== "")
            writer.tag(2, WireType.LengthDelimited).string(message.kind);
        /* octelium.api.main.meta.v1.Metadata metadata = 3; */
        if (message.metadata)
            Metadata.internalBinaryWrite(message.metadata, writer.tag(3, WireType.LengthDelimited).fork(), options).join();
        /* octelium.api.main.user.v1.Namespace.Spec spec = 4; */
        if (message.spec)
            Namespace_Spec.internalBinaryWrite(message.spec, writer.tag(4, WireType.LengthDelimited).fork(), options).join();
        /* octelium.api.main.user.v1.Namespace.Status status = 5; */
        if (message.status)
            Namespace_Status.internalBinaryWrite(message.status, writer.tag(5, WireType.LengthDelimited).fork(), options).join();
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message octelium.api.main.user.v1.Namespace
 */
export const Namespace = new Namespace$Type();
// ...
export const Namespace_Spec = new Namespace_Spec$Type();
// ...
export const Namespace_Status = new Namespace_Status$Type();
// ...
export const NamespaceList = new NamespaceList$Type();